:wedding: set pipx default version
installing llm@0.12 via pipx fails;
pipx installed via brew defaults to python@3.12, but aiohttp, which is somehow required
by a dep, is not compatible with python@3.12.

since only thing i'm using pipx for (right now) is this tooling, just globally
set the default for all envs to be a compatible version (in this case: python@3.11).